requirements.txt
ultralytics
streamlit
opencv-python-headless
Pillow
roboflow
inference-sdk
